Strangler extraction #16: SD-card op layer -> zxnu_sdcard_ops.py
Move the SD-card tab operation layer (~2,660 lines across four blocks
interleaved with the NextSync ops extracted in #15) out of
MainWindow.__init__ as four builders called at their historical
positions:

* build_sdcard_utils - is_hdfmonkey_present, image history, the
  transfer/hdfmonkey/load-image-hint animations, load_image, the three
  log windows, the new-folder chain, select_image,
  download_nextzxos_image, usage gauge + writability checks,
  execute_hdf_monkey and the shell-command helpers.
* build_image_edit_ops - image deletion (confirm dialog + worker) and
  in-image rename (task + dialog).
* build_local_explorer_ops - the SD tab local-pane ops: create-dir/
  rename helpers (shared with the NextSync classic explorer), context
  menu, watcher-suspended delete, zip/unzip, import.
* build_transfer_clipboard_ops - image <-> disk transfers, the shared
  cross-explorer clipboard (OS-clipboard synced), in-image copy/paste,
  remote zip/unzip, the image context menu and the SdCardExplorerPane
  delegate wrappers.

New seam piece: the module globals right_disk_image_explorer_content /
right_disk_image_selected_files stay in zxnu_main (reassigned by
closures still there) - the moved code reads them via the
_right_disk_content getter and writes them via _set_right_disk_content /
_set_right_disk_selected setter closures defined at the first call site;
the original global declarations are dropped in the token pass.

Also drop QDialog from the disclaimer closure lazy import (third
occurrence of the extraction-surfaces-latent-F811 pattern).
zxnu_main.py 7,513 -> 4,997 lines - the monolith is now 81% smaller than
the original 26k. Tripwired in test_pane_imports.py; __init__-wide
free-var check back to zero unresolved names.
